feat: add --expect filename option (#30)
Added the `--expect` option to specify the expected test statuses in a JSON file. Each object key is a suite name, and individual string keys and values are the test names and statuses. For example, the following spec file has a suite with 2 tests, first should pass, and the second is expected to fail:

```js
describe('1 passing 1 failing', () => {
  it('passes', () => { ... })

  it('fails', () => {
    expect(true).to.be.false
  })
})
```

You can list the expected test results in a JSON file, for example, see "cypress/failing/expected.json" 

```json
{
  "1 passing 1 failing": {
    "passes": "passed",
    "fails": "fail"
  }
}
```

You can if the tests really behave this way by running:

```shell
$ npx cypress-expect run --expect ./cypress/failing/expected.json
```

**Tip:** you do not have to list the passing tests. Every test not listed in the expected JSON is assumed "passing" by default.

**Tip 2:** you do not have to remember the precise name of each test status, the JSON file can use synonyms, like `passed`, `pass`, `passing`, `pass`.